simplify subWindow-handling
This changes the way sub-windows are opened in lightweightMode for some reasons. I think, it is complicated and not practical to add a custom navigationBar / back button to a sub-window. You can still do this (if you want) or use the ActionBar, just as you like. Instead, you can now open as many sub-windows as you want on a tab and close them with the native back-button (same UX as heavyweightMode).